Avoiding stack allocation
Fixes #4

Using into_iter was a "prettier" way to handle the dereferencing. I did
not think about the ramificiations of this on the stack, as I didn't
expecct the compiler to copy the word list array to the stack. This
would only have contained the pointers to the string data, so I'm
surprised it caused a stack overflow on any platform.

This change, however, should yield no copies of the array hitting the
stack and instead only copy the data while executing hte loop's contents.